fix(tests): update WS test helpers to unwrap WsEnvelope post-login
Introduces `WsSession` in `tests/common/mod.rs` that mirrors the
production session pattern — tracks an HMAC key derived from the
`SessionToken=` embedded in `InitConnectionSuccess`, seals outgoing
frames with `seal_ws_frame`, and opens incoming frames with
`open_ws_frame`.  All 18 integration-test files updated to use
`WsSession` instead of the bare `WebSocketStream`.

Also fixes:
- `WsHmacError: std::error::Error` impl missing in `common/src/crypto.rs`
- `request_id: i64` vs `u32` mismatch in `database/jobs.rs` test closure
- Unused `PyBytes`/`PyDict` imports in `plugins/mod.rs`; bare usages
  inside `#[cfg(test)]` switched to fully-qualified `pyo3::types::…`

Closes red-cell-c2-g2c7j
Closes red-cell-c2-hn9eb

+++ teamserver/src/database/jobs.rs
@@ -562,7 +562,7 @@ mod tests {
         let db = Database::connect_in_memory().await.expect("db");
         seed_agents(&db, &[101]).await;
         let repo = db.agent_responses();
-        let make_record = |request_id: i64, output: &str, received_at: &str| AgentResponseRecord {
+        let make_record = |request_id: u32, output: &str, received_at: &str| AgentResponseRecord {
             id: None,
             agent_id: 101,
             command_id: 21,
